% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseRNAseqPredictions.R
\name{analyseRNAseqPredictions}
\alias{analyseRNAseqPredictions}
\title{Calculate fold change to measure RNA-seq mutation effects}
\usage{
analyseRNAseqPredictions(
  predictions.dir,
  metadataWT,
  metadataMT,
  regions.gr = NULL,
  bedgraph.gz = FALSE,
  track_nums = NULL
)
}
\arguments{
\item{predictions.dir}{character. The directory containing the prediction files (BedGraph format) and the two metadata files.}

\item{metadataWT}{data.frame. The metadata for the wild-type sequences.}

\item{metadataMT}{data.frame. The metadata for the mutant sequences.}

\item{regions.gr}{GRanges. Optional GRanges object representing specific genomic regions of interest to calculate local fold change. Default is NULL.}

\item{bedgraph.gz}{logical. Default is FALSE. If TRUE, the function searches for gzipped files (.bedgraph.gz).}

\item{track_nums}{numeric. Optional vector of track numbers to analyze. If specified, only tracks with track_num in track_nums are processed. Default is NULL.}
}
\value{
A data.table containing the mutation metadata with appended fold change scores (gFC_x, min_gFC_x) for all tracks.
}
\description{
This function calculates fold change scores to measure the effect of mutations on gene/transcript expression levels (RNA-seq).
It reads predicted 1D track values (BedGraph files) from a specified directory, processes them based on WT and MT metadatas,
and computes global and optional local fold changes.
}
