% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvDNA.R
\name{cnvDNA}
\alias{cnvDNA}
\title{create copy number variations of sequences}
\usage{
cnvDNA(DNA.string, window, cnv.start, width, nb_copy, fpos = "center")
}
\arguments{
\item{DNA.string}{DNA input as DNAstring object (e.g chromosome sequence)}

\item{window}{output size in bp.}

\item{cnv.start}{start of the CNV in bp.}

\item{width}{width of the CNV in bp.}

\item{fpos}{final position in bp of the CNV according to the output sequence. By default fpos = "center" to center the CNV on the output sequence.}
}
\value{
DNAstring
}
\description{
From DNA sequence (e.g. chromosome), the function create X copy of a DNA segment (e.g. deletion, duplication...).
}
\examples{
DNA.seq = Biostrings::DNAString("AAAAAAAAAAAAAAGTTGCCCCCCCCCCCCCC")
DNA.seq

#duplication of "T" nucleotides (16:17)
#output window of 20 nucletodides with de duplication start at the center (i.e 11th bp)
cnvDNA(DNA.string = DNA.seq, window = 20,
    cnv.start = 16, width = 2, nb_copy = 2, fpos = "center")

#output window of 20 nucletodides with de duplication start at 6bp
cnvDNA(DNA.string = DNA.seq, window = 20,
    cnv.start = 16, width = 2, nb_copy = 2, fpos = 6)

#deletion of "T" nucleotides (16:17)
cnvDNA(DNA.string = DNA.seq, window = 20,
    cnv.start = 16, width = 2, nb_copy = 0)

}
