% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvDNA2.R
\name{cnvDNA2}
\alias{cnvDNA2}
\title{create copy number variations of sequences}
\usage{
cnvDNA2(dna.string, cnv.start, cnv.stop, nb_copy)
}
\arguments{
\item{dna.string}{DNA input as DNAstring object (e.g chromosome sequence)}

\item{cnv.start}{start of the CNV in bp.}

\item{cnv.stop}{stop of the CNV in bp.}

\item{nb_copy}{final number of copy of the CNV (0 copy for deletion).}
}
\value{
DNAstring
}
\description{
From DNA sequence (e.g. chromosome), the function create X copy of a DNA segment (e.g. deletion, duplication...).
}
\examples{
DNA.seq = Biostrings::DNAString("AAAAAAAAAAAAAAGTTGCCCCCCCCCCCCCC")
DNA.seq

#duplication of "T" nucleotides (16:17)
cnvDNA2(dna.string = DNA.seq, cnv.start = 16, cnv.stop = 17, nb_copy = 2)

#deletion of "T" nucleotides (16:17)
cnvDNA2(dna.string = DNA.seq, cnv.start = 16, cnv.stop = 17, nb_copy = 0)

}
