% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delMAT.R
\name{delMAT}
\alias{delMAT}
\title{delete bin on HiC matrix}
\usage{
delMAT(matrix, bin.width, start, stop, limits = "nearest")
}
\arguments{
\item{matrix}{interaction matrix for only one chromosome as \code{dgCMatrix} or \code{matrix} object.}

\item{bin.width}{bin width of the matrix.}

\item{start}{one or more positions of deletion start in bp.}

\item{stop}{one or more positions of deletion stop/end in bp.}

\item{limits}{"nearest" to removed bins according to the nearest bin boundaries (default), max to removed bins involved in deletion, min to removed bins fully involved in deletion (see example and details).}
}
\value{
DNAstring
}
\description{
From DNA position of a deletion, the function remove the bins concern by the deletion
}
\details{
When a deletion start or end is positioned in a middle of a bin (i.e. not at the boundary), 3 solutions are possible:
-remove the bin from the matrix (limits = "max"),
-keep the bin in the matrix (limits = "min"),
-choose the nearest bin boundary (limits = "nearest").
}
\examples{
matrix = sapply(1:5, function(x){rep(x, 5)}) ; matrix
bin.width = 10e3
start = 24e3
stop = 44e3

#round deletion to range 20e3 and 40e3
delMAT(matrix, bin.width, start, stop, limits = "nearest")

#round deletion to range 20e3 and 50e3
delMAT(matrix, bin.width, start, stop, limits = "max")

#round deletion to range 30e3 and 40e3
delMAT(matrix, bin.width, start, stop, limits = "min")

}
