% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insDNA2.R
\name{insDNA2}
\alias{insDNA2}
\title{insert a DNA segment into another DNAsegment}
\usage{
insDNA2(dna.string1, dna.string2, ins.start, rev = FALSE)
}
\arguments{
\item{dna.string1}{DNA input as DNAstring object to be inserted.}

\item{dna.string2}{DNA input as DNAstring object (e.g chromosome sequence) where the insertion will take place.}

\item{ins.start}{position in bp of the insertion on DNA.string2}

\item{rev}{whether or not to reverse the inserted sequence, default = FALSE.}
}
\value{
DNAstring
}
\description{
to do
}
\examples{
DNA.seq1 = Biostrings::DNAString("GATG")
DNA.seq2 = Biostrings::DNAString("AAAAACCCCCTTTTTAAAAA")

#insertion startint at bp number 11:
insDNA2(dna.string1 = DNA.seq1, dna.string2 = DNA.seq2, ins.start = 11, rev = FALSE)

#idem but reversed
insDNA2(dna.string1 = DNA.seq1, dna.string2 = DNA.seq2, ins.start = 11, rev = TRUE)

}
