% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invDNA.R
\name{invDNA}
\alias{invDNA}
\title{reverse a DNA segment (inversion)}
\usage{
invDNA(dna.string, start, stop)
}
\arguments{
\item{dna.string}{DNA input as DNAstring object (e.g chromosome sequence)}

\item{start}{position in bp of the first nucleotide to reverse}

\item{stop}{position in bp of the last nucleotide to reverse}
}
\value{
DNAstring
}
\description{
to do
}
\examples{
DNA.seq = Biostrings::DNAString("AAAAACCCCCAAAAACCCCC")

# inversion of CCCCCTTTTT nucleotides:
invDNA(dna.string = DNA.seq, start = 6, stop = 15)

}
