% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matZoomify.R
\name{matZoomify}
\alias{matZoomify}
\title{Decrease the resolution of an interaction matrix}
\usage{
matZoomify(matrix, old_bin_width, new_bin_width, verbose = TRUE)
}
\arguments{
\item{matrix}{\code{dgCMatrix} or \code{matrix} object representing intra-chromosomal interactions.}

\item{old_bin_width}{Bin width of the input matrix in base pairs (i.e. original resolution).}

\item{new_bin_width}{Target bin width in base pairs (i.e. new resolution). Must be a multiple of \code{old_bin_width}.}

\item{verbose}{Logical. Whether or not to print messages. Default = \code{TRUE}.}
}
\value{
An upper triangular \code{dgCMatrix} object with the decreased resolution.
}
\description{
\code{matZoomify} decreases the resolution of a Hi-C interaction matrix by binning.
For example, converting a matrix from a 50kb resolution to a 100kb resolution.
}
\details{
The input matrix is first converted to a sparse \code{CsparseMatrix} and restricted to its upper triangular part (including the diagonal).
Then, bins are aggregated by summing the interaction counts according to the ratio of \code{new_bin_width} to \code{old_bin_width}.
If the original number of bins is not a multiple of the reduction factor, a final smaller bin is created at the end to contain the remaining bins.
The resulting matrix is returned as an upper triangular \code{dgCMatrix} (sparse Matrix).
}
\examples{
# Decrease resolution of HCT116 chromosome 19 matrix from 50kb to 100kb
mat_100kb <- matZoomify(mat_HCT116_chr19_50kb,
                        old_bin_width = 50e3,
                        new_bin_width = 100e3)

}
