% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataMT.R
\name{metadataMT}
\alias{metadataMT}
\title{Mutated metadata dataframe generation}
\usage{
metadataMT(metadataWT, mutated.width, rep = 1)
}
\arguments{
\item{metadataWT}{\if{html}{\out{<dataFrame>}} : a dataframe generate by createMetadaWT}

\item{mutated.width}{\if{html}{\out{<numeric>}} : size of the adjacent mutated window}

\item{rep}{\if{html}{\out{<numeric>}} : the number of repetition for each mutations.}
}
\value{
dataframe
}
\description{
The main goal of this function is to create MT metadata to create fasta files for each mutation.
For each 1Mb WT sequences, this function return all the adjacent mutation positions on the mutated window.
}
\details{
This function return a dataframe with metadatas relate to Mutated sequences used to run Orca predictions :
Column description:
\itemize{
\item chr \if{html}{\out{<numeric>}} : the number of the chromosome
\item start.window \if{html}{\out{<numeric>}} : start of the window that will contain the mutations
\item stop.window \if{html}{\out{<numeric>}} : stop of the window that will contain the mutations
\item start.pred \if{html}{\out{<numeric>}} : the first position of the 1Mb sequence
\item stop.pred \if{html}{\out{<numeric>}} : the last position of the 1Mb sequence
\item start.mut \if{html}{\out{<numeric>}} : start of mutated window
\item stop.mut \if{html}{\out{<numeric>}} : stop of the mutated window
\item ID \if{html}{\out{<character>}} : the name of the mutated type sequence "MT_x_y_z.fa" with: x the number of the related WildType sequence, y an incremental number for each different mutated window, z an incremental number for each repetition
}
}
