% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataWT_customWindow.R
\name{metadataWT_customWindow}
\alias{metadataWT_customWindow}
\title{Wild type metadata dataframe generation for specified regions}
\usage{
metadataWT_customWindow(
  DNAstring,
  DNAstring_name,
  start.windows,
  stop.windows,
  max.window.size,
  model_HFF,
  model_ESC
)
}
\arguments{
\item{DNAstring}{DNAString : the DNA sequence (e.g. chromosome).}

\item{DNAstring_name}{character : the sequence name (e.g 1 or chr1...)}

\item{start.windows}{numeric : vector of start positions of the windows that will contain the mutations.}

\item{stop.windows}{numeric : vector of stop positions of the windows that will contain the mutations.}

\item{max.window.size}{numeric : juxtaposed window will be reduce. If the width of the window is greater to max.window.size, the widow will be split. Must be smaller than 1Mb.}

\item{model_HFF}{\if{html}{\out{<boolean>}} : use HFF training model for Orca predictions}

\item{model_ESC}{\if{html}{\out{<boolean>}} : use ESC training model for Orca predictions}
}
\value{
dataframe
}
\description{
The main goal of this function is to create WT metatdata to run Orca 1Mb predictions only for the selected mutated windows.
The purpose of this function is to reduce the number of WT sequences needed to compare MT sequences. To do so, one WT sequence will be used as control for many MT sequences (number depending on how many mutation will be performed per mutated window).
For each 1Mb WT sequences, this function return the window that will be subject to mutations.
Those metadata will be used by createMetadaMT function to create the metadata of the mutant sequences (MT).
}
\details{
This function return a dataframe with metadatas relate to WildType sequences used to run Orca predictions :
Column description:
\itemize{
\item chr \if{html}{\out{<numeric>}} : the number of the chromosome
\item start.window \if{html}{\out{<numeric>}} : start of the window that will contain the mutations
\item stop.window \if{html}{\out{<numeric>}} : stop of the window that will contain the mutations
\item start.pred \if{html}{\out{<numeric>}} : the first position of the 1Mb sequence
\item stop.pred \if{html}{\out{<numeric>}} : the last position of the 1Mb sequence
\item ID \if{html}{\out{<character>}} : the name of the wild type sequence "WT_x.fa" with x an incremental number
\item A,C,T,G \if{html}{\out{<numeric>}} : 4 column with the frequences of each nucleotides on the 1Mb sequence.
\item model = 1000000 : ORCA parameter cf ORCA documentation
\item scale = 1000000 : ORCA parameter cf ORCA documentation
\item wpos  500000 : ORCA parameter cf ORCA documentation
\item mpos  500000 : ORCA parameter cf ORCA documentation
\item model_HFF <0 or 1> : ORCA parameter. predict HFF matrices if 1
\item model_ESC <0 or 1> : ORCA parameter. predict ESC matrices if 1
}
}
