% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataWT_customWindow_Alphagenome.R
\name{metadataWT_customWindow_Alphagenome}
\alias{metadataWT_customWindow_Alphagenome}
\title{Wild type metadata dataframe generation for specified regions with AlphaGenome}
\usage{
metadataWT_customWindow_Alphagenome(
  DNAstring,
  DNAstring_name,
  start.windows,
  stop.windows,
  max.window.size,
  organism,
  ontology
)
}
\arguments{
\item{DNAstring}{DNAString : the DNA sequence (e.g. chromosome).}

\item{DNAstring_name}{character : the sequence name (e.g 1 or chr1...)}

\item{start.windows}{numeric : vector of start positions of the windows that will contain the mutations.}

\item{stop.windows}{numeric : vector of stop positions of the windows that will contain the mutations.}

\item{max.window.size}{numeric : juxtaposed window will be reduce. If the width of the window is greater to max.window.size, the widow will be split. Must be smaller than 1048576.}

\item{organism}{character : organism for AlphaGenome prediction (e.g., "human" or "mouse")}

\item{ontology_term}{character : ontology term for AlphaGenome prediction}
}
\value{
dataframe
}
\description{
The main goal of this function is to create WT metatdata to run AlphaGenome 1048576 bp predictions only for the selected mutated windows.
Those metadata will be used by createMetadaMT function to create the metadata of the mutant sequences (MT).
}
\details{
This function return a dataframe with metadatas relate to WildType sequences used to run AlphaGenome predictions :
Column description:
\itemize{
\item chr \if{html}{\out{<numeric>}} : the number of the chromosome
\item start.window \if{html}{\out{<numeric>}} : start of the window that will contain the mutations
\item stop.window \if{html}{\out{<numeric>}} : stop of the window that will contain the mutations
\item start.pred \if{html}{\out{<numeric>}} : the first position of the 1048576 bp sequence
\item stop.pred \if{html}{\out{<numeric>}} : the last position of the 1048576 bp sequence
\item ID \if{html}{\out{<character>}} : the name of the wild type sequence "WT_x.fa" with x an incremental number
\item A,C,T,G \if{html}{\out{<numeric>}} : 4 column with the frequences of each nucleotides on the 1048576 bp sequence.
\item scale = 1048576 : ALPHAGENOME parameter
\item organism \if{html}{\out{<character>}} : ALPHAGENOME parameter
\item ontology \if{html}{\out{<character>}} : ALPHAGENOME parameter
}
}
