% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataWT_seqWide.R
\name{metadataWT_seqWide}
\alias{metadataWT_seqWide}
\title{Wild type metadata dataframe generation sequence wide.}
\usage{
metadataWT_seqWide(
  DNAstring,
  DNAstring_name,
  window.size,
  model_HFF,
  model_ESC
)
}
\arguments{
\item{DNAstring}{DNAString : the DNA sequence (e.g. chromosome).}

\item{DNAstring_name}{character : the sequence name (e.g "chr1")}

\item{window.size}{numeric : size of adjacent windows in bp which indicate mutations area at the center of the 1Mb sequences. Must be smaller than 1Mb (see details)}

\item{model_HFF}{\if{html}{\out{<boolean>}} : use HFF training model for Orca predictions}

\item{model_ESC}{\if{html}{\out{<boolean>}} : use ESC training model for Orca predictions}
}
\value{
dataframe
}
\description{
The main goal of this function is to create WT metatdata to run Orca 1Mb predictions all along the sequence provided.
The purpose of this function is to reduce the number of WT sequences needed to compare MT sequences. To do so, one WT sequence will be used as control for many MT sequences (number depending on how many mutation will be performed per window.size).
For each 1Mb WT sequences, this function return the window that will be subject to mutations, ensuring that these mutated windows are adjacent along the sequence
Those metadata will be used by createMetadaMT function to create the metadata of the mutant sequences (MT).
}
\details{
The window.size parameter is used to define the start and stop of the mutated area for each 1Mb sequences. This window is positioned at the center of the 1Mb sequences that will be predicted.
For example: For the first 1Mb sequence (1 to 1e6 bp), if window = 256000, the fist mutation will start at 372001 (1e6 / 2 - 256e3 / 2 + 1) and the last mutation will start at 628000 (1e6 / 2 - 256e3 / 2).

This function return a dataframe with metadatas relate to WildType sequences used to run Orca predictions :
Column description:
\itemize{
\item chr \if{html}{\out{<numeric>}} : the name of the sequence
\item start.window \if{html}{\out{<numeric>}} : start of the window that will contain the mutations
\item stop.window \if{html}{\out{<numeric>}} : stop of the window that will contain the mutations
\item start.pred \if{html}{\out{<numeric>}} : the first position of the 1Mb sequence
\item stop.pred \if{html}{\out{<numeric>}} : the last position of the 1Mb sequence
\item ID \if{html}{\out{<character>}} : the name of the wild type sequence "WT_x.fa" with x an incremental number
\item A,C,T,G \if{html}{\out{<numeric>}} : 4 column with the frequencies of each polynucleotides on the 1Mb sequence.
\item model = 1000000 : ORCA parameter cf ORCA documentation
\item scale = 1000000 : ORCA parameter cf ORCA documentation
\item wpos  500000 : ORCA parameter cf ORCA documentation
\item mpos  500000 : ORCA parameter cf ORCA documentation
\item model_HFF <0 or 1> : ORCA parameter. predict HFF matrices if 1
\item model_ESC <0 or 1> : ORCA parameter. predict ESC matrices if 1
}
}
