% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIC.R, R/SIC2.R, R/matFC.R
\name{SIC}
\alias{SIC}
\title{Structural impact score between 2 matrices area}
\usage{
SIC(mat1, mat2)

SIC(mat1, mat2)

SIC(mat1, mat2)
}
\arguments{
\item{mat1, mat2}{input matrices as \code{dgCMatrix} or \code{matrix} object for only one chromosome.}

\item{bin.width}{Bin width of the matrix in base pair.}

\item{vp.start}{Start of the view point in base pair.}

\item{vp.stop}{Stop/end of the view point in base pair. Default is NULL to only use the bin where vp.start is located (ie: vp.stop = vp.start + bin.width).}

\item{start, stop}{Area in bp of the chromosome to compute impact score. Default is NULL to use the entire chromosome (i.e. entire matrix).}

\item{verbose}{if TRUE, show information when view point does not match a bin annotations.}

\item{mutated.mat, wildtype.mat}{mutated (query) or wildtype (control, subject) matrices (observed or observed/expected matrices) as \code{dgCMatrix} or \code{matrix} object for only one chromosome.}
}
\value{
integer

integer

integer
}
\description{
Measurement described in Zhou 2022 use to compute the average absolute log fold change of interactions between 2 matrices.
Fold change can be compute:
\itemize{
\item from one bin to all bins in the matrix (using view point parameters) : useful to measure the impact on the interactions of a specific bin (e.g. bin containing a mutation) or bins (e.g. bins from a TAD) with all other bins in the matrix.
\item to a specific area of the matrix (using start/stop parameters) : useful to measure the impact between a TAD (the view point) on another TAD...
}

Measurement described in Zhou 2022 use to compute the average absolute log fold change of interactions between each bin of 2 matrices.
Fold change is computed for each bin of the matrices provided

Fold change is computed for each bin of the matrices provided and the absolute log2 mean fold change is returned.
}
\details{
The function takes as input 2 matrices (mutated and wildtype) and computes the average absolute log fold change between each bins (return one value for the entire matrices).
}
\examples{
# to do

# to do

# to do

}
