% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapDNA.R
\name{swapDNA}
\alias{swapDNA}
\title{DNA swapping within sequence}
\usage{
swapDNA(dna.string, center, window, origin.start, width, target.start)
}
\arguments{
\item{dna.string}{DNA input as DNAstring object}

\item{center}{center of output DNA window in bp.}

\item{window}{output size in bp (centered around `center`). E.g, for +/-0.5 Mb on either side of `center`, use 1 Mb.}

\item{origin.start}{start of the segment to be swapped in bp.}

\item{width}{width of the segment to be swapped in bp.}

\item{target.start}{start of the segment to be overwriting in bp.}
}
\value{
DNAstring
}
\description{
From DNA sequence (e.g. chromosome), the function copy a DNA segment which is moved to another position.
The positioning is done by overwriting the destination sequence of the same size.
}
\details{
Note that to simplify, DNA positions are 0 base!
For example, if origin.start = 1e6 (1Mb) and width = 1e3, the first nucleotide position will be 1,000,001bp and the last 1,001,000.
}
\examples{
dna.seq = Biostrings::DNAString("AAAAAAAAAATTTTTTTTTTCCCCCCCCCC")
dna.seq

#swapping of the first 10 nucleotides (1:10) in the last 10 nucleotides (21:30)
#output window of 20 nucletodides (i.e. 11:30)
swapDNA(dna.string = dna.seq, center = 20, window = 20,
    origin.start = 0, width = 10, target.start = 20)

}
