% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trcpDNA.R
\name{trcpDNA}
\alias{trcpDNA}
\title{create reciprocal translocation between two sequences}
\usage{
trcpDNA(DNA.string1, DNA.string2, window, breakpoint = NULL)
}
\arguments{
\item{DNA.string1}{DNA input as DNAstring object.}

\item{DNA.string2}{DNA input as DNAstring object.}

\item{window}{output size in bp. E.g, for +/-0.5 Mb of each sequences, use 1 Mb.}

\item{breakpoint}{position in bp of the breakpoint (i.e position where the change occurs) according to the exported sequence. Default is NULL to use the center of the window as breakpoint (see examples).}
}
\value{
DNAstring
}
\description{
From two DNA sequence (e.g. chromosome), the function fused a segment of the 2 sequences to create a new one.
}
\details{
By default the breakpoint of the output sequence is the middle of window. For example, if window is 10 bp, the breakpoint is between base pair number 5 and 6.
}
\examples{
DNA.seq1 = Biostrings::DNAString("AAAAAGGGGG")
DNA.seq2 = Biostrings::DNAString("CCCCCTTTTT")

# reciprocal translocation of the DNA.seq1 with DNA.seq1
trcpDNA(DNA.string1 = DNA.seq1, DNA.string2 = DNA.seq2, window = 10)

# reciprocal translocation of the DNA.seq1 with DNA.seq1, which sequence change at third bp (breakpoint = 3)
trcpDNA(DNA.string1 = DNA.seq1, DNA.string2 = DNA.seq2, window = 10, breakpoint = 3)

}
