% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFastaMT.R
\name{writeFastaMT}
\alias{writeFastaMT}
\title{metadata MT to fasta sequences files}
\usage{
writeFastaMT(DNAstring, metadataWT, metadataMT, workdir = "./", gzip = TRUE)
}
\arguments{
\item{DNAstring}{\if{html}{\out{<DNAString>}} : the sequence of the chromosome, it must be the same sequence that was use to generate the WildType metadata dataFrame (with genDFWT or genDFWTstep2)}

\item{metadataWT}{\if{html}{\out{<dataFrame>}} : a dataframe generated by metadataWT_seqWide or metadataWT_customWindow.}

\item{metadataMT}{\if{html}{\out{<dataFrame>}} : a dataframe generated by metadataMT.}

\item{workdir}{\if{html}{\out{<character>}} : the path to the folder that will contain a "fasta" directory with all the fasta sequences. The folder are created if they doesn't exist.}

\item{gzip}{\if{html}{\out{<boolean>}} : if TRUE, the fasta files are saved in gzip format, otherwise they are saved in plain text format.}
}
\value{
\if{html}{\out{<message>}} : a message indicating the number of fasta files saved and the output directory.
}
\description{
Generate and saved mutated sequences as fasta files.
The mutated sequences are generated by random calling of nucleotides based on nucleotide frequencies of the 1Mb sequences.
}
