% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFastaWT.R
\name{writeFastaWT}
\alias{writeFastaWT}
\title{metadata WT to fasta sequences files}
\usage{
writeFastaWT(DNAstring, metadataWT, workdir = "./", gzip = TRUE)
}
\arguments{
\item{DNAstring}{\if{html}{\out{<DNAString>}} : the sequence of the chromosome, it must be the same sequence that was use to generate the WildType metadata dataFrame (metadataWT_seqWide or metadataWT_customWindow).}

\item{metadataWT}{\if{html}{\out{<dataFrame>}} : a dataframe generate by metadataWT_seqWide or metadataWT_customWindow.}

\item{workdir}{\if{html}{\out{<character>}} : the working directory where the fasta files will be saved. Default is "./".}

\item{gzip}{\if{html}{\out{<logical>}} : if TRUE, the fasta files will be saved in gzip format. Default is TRUE.}
}
\value{
\if{html}{\out{<message>}} : a message indicating the number of fasta files saved and the output directory.
}
\description{
metadata WT to fasta sequences files
}
