% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeOrcaData.R
\name{writeOrcaData}
\alias{writeOrcaData}
\title{Wrapper for Orca data preparation}
\usage{
writeOrcaData(
  DNAstring,
  DNAstring_name,
  window.size = NULL,
  start.windows = NULL,
  stop.windows = NULL,
  max.window.size = NULL,
  model_HFF,
  model_ESC,
  mutated.width,
  workdir,
  gzip = FALSE,
  rep = 1
)
}
\arguments{
\item{DNAstring}{DNAString : the DNA sequence (e.g. chromosome).}

\item{DNAstring_name}{character : the sequence name (e.g 1 or chr1...)}

\item{window.size}{numeric : size of adjacent windows in bp which indicate mutations area at the center of the 1Mb sequences. Must be smaller than 1Mb (see details)}

\item{start.windows}{numeric : vector of start positions of the windows that will contain the mutations.}

\item{stop.windows}{numeric : vector of stop positions of the windows that will contain the mutations.}

\item{max.window.size}{numeric : juxtaposed window will be reduce. If the width of the window is greater to max.window.size, the widow will be split. Must be smaller than 1Mb.}

\item{model_HFF}{\if{html}{\out{<boolean>}} : use HFF training model for Orca predictions}

\item{model_ESC}{\if{html}{\out{<boolean>}} : use ESC training model for Orca predictions}

\item{mutated.width}{\if{html}{\out{<numeric>}} : size of the adjacent mutated window}

\item{workdir}{\if{html}{\out{<character>}} : the path to the folder that will contain a "fasta" directory with all the fasta sequences. The folder are created if they doesn't exist.}

\item{gzip}{\if{html}{\out{<boolean>}} : if TRUE, the fasta files are saved in gzip format, otherwise they are saved in plain text format.}

\item{rep}{\if{html}{\out{<numeric>}} : the number of repetition for each mutations.}
}
\value{
\if{html}{\out{<message>}} : a message indicating the number of fasta files saved and the output directory.
}
\description{
description
This function prepares the data for Orca by generating WildType and Mutated metadata, writing the corresponding fasta files, and saving the metadata as tsv files.
This function run the following functions:
\itemize{
\item metadataWT_customWindow or metadataWT_seqWide to generate WildType metadatas
\item metadataMT to generate Mutated metadatas
\item writeFastaWT to write WildType fasta files
\item writeFastaMT to write Mutated fasta files
}
}
