% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseOrcaPredictions.R
\name{analyseOrcaPredictions}
\alias{analyseOrcaPredictions}
\title{Calculate scores to measure mutation effects}
\usage{
analyseOrcaPredictions(
  predictions.dir,
  metadataWT,
  metadataMT,
  matrix.gz = FALSE,
  SIC = TRUE,
  corr = FALSE,
  DI = FALSE
)
}
\arguments{
\item{predictions.dir}{character. The directory containing the prediction files.}

\item{metadataWT}{data.frame. The metadata for the wild-type matrices.}

\item{metadataMT}{data.frame. The metadata for the mutant matrices.}

\item{matrix.gz}{logical. If TRUE, the function expects gzipped matrix files. Default is FALSE.}

\item{SIC}{logical. If TRUE, the function calculates the SIC score. Default is TRUE.}

\item{corr}{logical. If TRUE, the function calculates the correlation score. Default is FALSE.}

\item{DI}{logical. If TRUE, the function calculates the DI score. Default is FALSE.}
}
\value{
data.frame
}
\description{
This function calculates scores to measure the effect of mutations on chromatin structure based on WT matrices (\eqn{mat_{WT} = log{\frac{Observed}{Expected}}}) and their corresponding MT matrices (\eqn{mat_{MT} = log{\frac{Observed}{Expected}}}).
It reads the prediction files from a specified directory, processes them based on the provided metadatas, and computes the scores (see details).
}
\details{
The scores are calculated for each MT matrix against its corresponding wild-type matrix, based on the metadata provided.
The function calculates the following scores:
\itemize{
\item \strong{SIC}: The mean absolute log fold change between the MT and WT matrices. \deqn{\text{SIC} = mean(abs(log_{2}\frac{mat_{MT}}{mat_{WT}})) = \frac{1}{N} \sum_{(i,j) \in \Omega} \left| \log_2 \left( \frac{mat^{MT}_{i,j}}{mat^{WT}_{i,j}} \right) \right|}
\item \strong{corr}: The Pearson correlation coefficient between the MT and WT matrices. \deqn{\text{corr} = corr(mat_{MT}, mat_{WT})}
\item \strong{DI}: Currently not implemented, returns NA.
}

The function also handles cases where there are multiple mutations in the same region (i.e., replicates) and adds columns to indicate which mutation has the best score (highest for corr, lowest for SIC).
The results are returned as a data.frame that combines the metadata for the mutant matrices with the calculated scores.
The function supports both HFF and ESC models, and can handle gzipped matrix files.
}
