% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseOrcaPredictions.R
\name{analyseOrcaPredictions}
\alias{analyseOrcaPredictions}
\title{Calculate scores to measure mutation effects}
\usage{
analyseOrcaPredictions(
  predictions.dir,
  metadataWT,
  metadataMT,
  matrix.gz = FALSE,
  SIC = TRUE,
  corr = FALSE,
  DI = FALSE
)
}
\arguments{
\item{predictions.dir}{character. The directory containing the prediction files.}

\item{metadataWT}{data.frame. The metadata for the wild-type matrices.}

\item{metadataMT}{data.frame. The metadata for the mutant matrices.}

\item{matrix.gz}{logical. If TRUE, the function expects gzipped matrix files. Default is FALSE.}

\item{SIC}{logical. If TRUE, the function calculates the SIC score. Default is TRUE.}

\item{corr}{logical. If TRUE, the function calculates the correlation score. Default is FALSE.}

\item{DI}{logical. If TRUE, the function calculates the DI score. Default is FALSE.}
}
\value{
data.frame
}
\description{
This function calculates scores (see details) between WT matrices and their corresponding MT matrices.
It reads the prediction files from a specified directory, processes them based on the provided metadatas, and computes the scores for each mutant matrix against its corresponding wild-type matrix.
The function supports both HFF and ESC models, and can handle gzipped matrix files.
The function return the metadataMT with the scores added as new columns.
}
\details{
The function calculates the following scores:
\itemize{
\item \strong{SIC}: The mean absolute log fold change between the MT and WT matrices.
\item \strong{corr}: The Pearson correlation coefficient between the MT and WT matrices.
\item \strong{DI}: Currently not implemented, returns NA.
}
}
