% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvDNA.R
\name{cnvDNA}
\alias{cnvDNA}
\title{create copy number variations of sequences}
\usage{
cnvDNA(dna.string, start, stop, nb_copy)
}
\arguments{
\item{dna.string}{DNA input as DNAstring object (e.g chromosome sequence)}

\item{start}{start of the CNV in bp.}

\item{stop}{stop of the CNV in bp.}

\item{nb_copy}{final number of copy of the CNV (0 copy for deletion).}
}
\value{
DNAstring
}
\description{
From DNA sequence (e.g. chromosome), the function create X copy of a DNA segment (e.g. deletion, duplication...).
}
\examples{
DNA.seq = Biostrings::DNAString("AAAAAAAAAAAAAAGTTGCCCCCCCCCCCCCC")
DNA.seq

# duplication of "T" nucleotides (positions 16:17)
cnvDNA(dna.string = DNA.seq, start = 16, stop = 17, nb_copy = 2)

# deletion of "T" nucleotides (16:17)
cnvDNA(dna.string = DNA.seq, start = 16, stop = 17, nb_copy = 0)

}
