% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shufDNA.R
\name{shufDNA}
\alias{shufDNA}
\title{DNA shuffle}
\usage{
shufDNA(dna.string, start, stop, probability = c(0.25, 0.25, 0.25, 0.25))
}
\arguments{
\item{dna.string}{DNA input as DNAstring object}

\item{start}{start of the segment to shuffle in bp.}

\item{stop}{stop of the segment to shuffle in bp.}

\item{probability}{A,T,G and C nucleotides probabilities respectively. Default is c(0.25, 0.25, 0.25, 0.25) to randomly call each nucleotide. Use probability = "full" to perform randomization based on the nucleotide probability of the entire sequence supplied. Use the “partial” parameter to take into account only the probability of the segment to be randomized.}
}
\value{
DNAstring
}
\description{
Shuffle of a DNA segment with random or custom nucleotides calling.
}
\details{
The nucleotides calling can be based with differents nucleotides probabilities (see examples):
 1) the nucleotide frequency of the full sequence provided,
 2) the nucleotide frequency of the segment to shuffle,
 3) an equivalent frequency for each nucleotide (i.e random calling).
 4) custom probabilities of each nucleotides.
}
\examples{
dna.seq = Biostrings::DNAString("AAAAATTTTTGGGGGAAAAA")
dna.seq

# shuffle of T and G nucleotides (6:15) with random calling of nucleotides,
# i.e probability = (0.25, 0.25, 0.25, 0.25):
shufDNA(dna.string = dna.seq, start = 6, stop = 15)

# shuffle of T and G nucleotides (6:15) based on the probability of the entire sequence provided,
# i.e equivalent to probability = c(0.5, 0.25, 0.25, 0) in this example::
shufDNA(dna.string = dna.seq, start = 6, stop = 15, probability = "full")

# idem but based on the probability of the segment to shuffle,
# i.e equivalent to probability = c(0, 0.5, 0.5, 0) in this example:
shufDNA(dna.string = dna.seq, start = 6, stop = 15, probability = "partial")

# idem but with custom probabilities:
# probability = c(0.5, 0, 0, 0.5) for A,T,G and C respectively:
shufDNA(dna.string = dna.seq, start = 6, stop = 15, probability = c(0.5, 0, 0, 0.5))

}
