% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shufDNA.R
\name{shufDNA}
\alias{shufDNA}
\title{DNA shuffle within sequence}
\usage{
shufDNA(DNA.string, center, window, shuffle.start, width)
}
\arguments{
\item{DNA.string}{DNA input as DNAstring object}

\item{center}{center of output DNA window in bp.}

\item{window}{output size in bp (centered around `center`). For +/-0.5 Mb on either side of `center`, use 1 Mb.}

\item{shuffle.start}{start of the segment to be swapped in bp.}

\item{width}{width of the segment to be swapped in bp.}
}
\value{
DNAstring
}
\description{
From DNA sequence (e.g. chromosome), the function compute the nucleotide frequencies which is used to generate a random sequence of a DNA segment.
}
\details{
Note that fo simplify, DNA positions are 0 base!
For example, if shuffle.start = 1e6 (1Mb) and width = 1e3, the first nucleotide position will be 1,000,001bp and the last 1,001,000.
}
\examples{
DNA.seq = Biostrings::DNAString("AAAAAAAAAATTTTTTTTTTCCCCCCCCCC")
DNA.seq

#shuffle of the last 10 nucleotides (21:30)
#output window of 30 nucletodides (i.e. 1:30)
shufDNA(DNA.string = DNA.seq, center = 15, window = 30,
    shuffle.start = 20, width = 10)

}
