% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIC.R
\name{SIC}
\alias{SIC}
\title{Structural impact score between 2 matrices}
\usage{
SIC(
  mutated.mat,
  wildtype.mat,
  bin.width,
  vp.start,
  vp.stop = NULL,
  start = NULL,
  stop = NULL,
  verbose = TRUE
)
}
\arguments{
\item{mutated.mat, wildtype.mat}{mutated (query) or wiltype (control, subject) matrices as `dgCMatrix` or `matrix` object for only one chromosome.}

\item{bin.width}{Bin width of the matrix in base pair.}

\item{vp.start}{Start of the view point in base pair.}

\item{vp.stop}{Stop/end of the view point in base pair. Default is NULL to only use the bin where vp.start is located.}

\item{start, stop}{Area in bp of the chromosome to be analyzed. Default is NULL to use the entire chromosome (i.e. entire matrix).}

\item{verbose}{if TRUE, show information when view point does not match a bin.}
}
\value{
integer
}
\description{
Measurement described in Zhou 2022 that calculates the average absolute log fold change of interactions between bin(s) (view point) and all other bins in specified window (start and stop).
}
\examples{
# to do

}
